requirements.txt
fastapi
uvicorn[standard]
pydantic
transformers
sentence-transformers
chromadb
vllm
python-dotenv
pandas
pyyaml
numpy
scikit-learn
torch
bitsandbytes
